try to move cidr utils into internals package
