+++ checkstyle/import-control.xml
@@ -259,7 +259,9 @@
 
     <subpackage name="authorizer">
       <allow pkg="org.apache.kafka.server.authorizer" />
-      <allow pkg="org.apache.commons.net.util" />
+      <subpackage name="internals">
+        <allow pkg="org.apache.commons.net.util" />
+      </subpackage>
     </subpackage>
 
     <subpackage name="telemetry">
