+++ metadata/src/main/java/org/apache/kafka/metadata/authorizer/StandardAuthorizerData.java
@@ -32,7 +32,7 @@ import org.apache.kafka.common.utils.internals.SecurityUtils;
 import org.apache.kafka.server.authorizer.Action;
 import org.apache.kafka.server.authorizer.AuthorizableRequestContext;
 import org.apache.kafka.server.authorizer.AuthorizationResult;
-import org.apache.kafka.server.authorizer.CidrUtils;
+import org.apache.kafka.server.authorizer.internals.CidrUtils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
